:truck: move app context provider to designated context folder - tested and passing
